pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "NLSS-PyDTS"
version = "2.0.0"
authors = [
  { name="Next Level Software Studio", email="seu@email.com" },
]
classifiers = [
  "Operating System :: Microsoft :: Windows :: Windows 11",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python",
  "License :: OSI Approved :: MIT License",
  "Topic :: System :: Networking",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Intended Audience :: Developers",
  "Natural Language :: English",
]
description = "A Next Level Software Studio project."
readme = "README.md"
requires-python = ">=3.14"
dependencies = ["NLSS-PyTools"]

[project.urls]
"Homepage" = "https://github.com/NextLevelSoftwareStudio/NLSS-PyDTS"

[tool.hatch.build.targets.wheel]
packages = ["src/pydts"]

[project.scripts]
pydts = "pydts.sys.cli:main"